Replace dataset cloning with an immutable base
Instead of cloning the (huge) dataset a whenever we need to set up a new
graph, we instead store a single copy of the dataset, and deep-freeze it
using IceNine. An ObjectDataset is then assigned to each Qernel object
which takes care of setting any values required for calculating a
scenario.

The ObjectDatasets are then replaced at the start of the next request.

No significant performance improvement, but there seems to be less load
on GC resulting in fewer requests "spiking" to 2-3x the normal response
time.